:fire: Remove dependency on dockertest
Though useful dockertest pulls in a crapton of dependencies which makes
our go.mod look pretty scary. The convenience of it starting a container
for you isn't worth the baggage.

Instead you're now expected to have an MQTT broker up and running
yourself and use `MQTT_ADDRESS` to point the tests at it if it doesn't
match the default in testutils.