Fix disconnect behaviour
We were previously returning nil from DisconnectPacketBuilder, which
results in no disconnect being sent. That was done in order to ensure
the broker always sends the LWT, as that's not sent on a regular
graceful disconnect.

This updates the packet builder to instead send a packet with reason
code 0x04, indicating to the broker that the client is disconnecting and
requiring it to send the LWT.